"Add user id field to task model and update routes to include user id for task creation, retrieval, and deletion"
This commit adds the `user` field to the `taskSchema` in the `task.model.js` file, which is a reference to the `User` model. The `user` field is required and is used to store the id of the user who created the task.

The `tasks.controller.js` file has also been updated to include the `user` field when creating and retrieving tasks. The `getTasks` and `createTask` functions now include the `user` field in the query and the `getTask` function now includes the `populate("user")` method to include the user information in the task object.

The `task.routes.js` file has been updated to include the `:id` parameter in the `delete` route, so that the correct task can be deleted based on the user id.